A1/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
numpy==1.24.4
opencv-python==4.8.0.76
Pillow==10.0.0
requests==2.31.0
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4


